chore: moved pagination hook to the context he is used
